+++ CHANGELOG.rst
@@ -7,7 +7,7 @@ The format is based on `Keep a Changelog <https://keepachangelog.com/en>`_
 and this project adheres to `Semantic Versioning <https://semver.org/spec/v2.0.0.html>`_.
 
 [Unreleased]
--------------
+------------
 
 Changed
 ^^^^^^^
@@ -18,4 +18,4 @@ Fixed
 Added
 ^^^^^
 
-- Project Created
+- Path class for camera moves